Add Protocol Footer support to project
Note that this uses the Footer as specced in Protocol, which is different
from the one on mozilla.org, in terms of how the social media links are
handled. We need the mozilla.org/Bedrock footer backporting to Protocol to
get that.

The Footer is configured as a Setting, because we only want one per site.
The Setting model has StreamFields to make the content flexible for the
current implementation of the Footer HTML, with an eye on the better/newer
Bedrock-specced Footer component.