Add newsletter-form block
This changeset adds the most complex block yet to birdbox: the Mozilla newsletter form.

The form submission is entirely JS-driven, and it uses JS lifted from Bedrock.

For now, no L10N is supported, but must follow on.

We also now support setting dimensions on images, which makes them easier to use.

The list of countries used is fixed for a temporary period - we'll drop in
a live list from product-details soon.