Lay in initial management command + data to bootstrap a default Mozilla footer
The data we're loading is based on the JSON format of the data stored in the DB
when you manually populate it via Wagtail.

Needed a little tweak to a couple of blocks to allow for the "rel" attr, which
we use for Mastodon verification and more.

Tests still to come