Main push at initial Docker setup
Uses sqlite and local/in-container file storage for now. Backing services will
come in a subsequent commit.

Supports building dev and relase images (+ related images) in the same way
we do for Bedrock, but with not as many variations because we don't need them

Includes docker-compose config to spin up dev/"app", release or release-local
versions of the build.