Add support for Google Cloud Storage file uploads
This commit adds support for env-driven config for GCS, using django-storages.
It's primarily designed for use in deployed environments, but can be used
locally (eg to develop or debug things that need cloud storage) by
adding GCP credentials to the local-credentials dir and referencing them
in the docker/envfiles/local.env file, along with other details needed to
access the bucket.