Support an allowlist of page types that may be added to a site (#222)
Because we'll be using forks of mozmeao/birdbox for multiple sites over time
we are likely to end up with additional Page models which are relevant to
only one (or a few) websites.

This changeset adds support for allowlist of `app_name.PageModelName` specs
for which page types are allowed to be added. By default, all page types
are allowed.

Resolves #213