Add support for custom social-sharing images
Added via wagtailmetadata package, which gives us OG and sharing tags for
FB, Twitter, etc via a simple templatetag and an additonal model field

We can't add the templatetag to the base template, because it will cause
pages like 404.html to blwo up, because they don't have `self` in their
context. So, instead, the tag has to be called on actual page templates,
not a base.