Add a Theme Color select (#276)
* Initial pass at adding a Theme Color select
... instead of picking an arbitrary color and a light or dark theme to go with it

This changeset adds a new ThemeColorBlock option, which is basically
a <select> where the value is a CSS class string comprising a theme class
(mzp-l-dark|light) and also one of 10 color classes, drawn from the protocol
color palette.

This new ThemeColorBlock is used in the HeroBlock and the ColumnBlock, to
replace the two places we used a ColorBlock + a ThemeBlock. It's a breaking
change, but it's one that we can fixup after deployment with likely no one
noticing.

This work is not complete, though. Outstanding points to fix:

1) The hero text area used to have a drop shadow on it which was based on
the color from the ColorBlock, which is now not at all available to the CSS.
This means we can remove the inline CSS (yay) but we'll need a new approach
to make the drop-shadow work.

2) The colors themselves selected for the mozorg and firefox themes are just
proofs of concept and could do with some tuning. The only two that MUST
remain (or have very similar ones) for consistency on future.m.o are
color-purple-10 and color-green-20

* Update changelog

* color scheme updates

* Flip the ThemedColorBlock options back to placeholder labels, so that we can
dynamically change them, based on the theme in use.

This because light-01 in Mozorg might be Color X, but in Firefox theme it's Y.

This work gives us stable keys to replace with human-readable names instead,
in the next commmit

* Add a hook to run some JS to replace the labels of the ThemedColorBlock select...
with approriate human-friendly labels that fit the active theme for the site.

* Update comments related to ThemedColorBlock and dynamic labels

---------

Co-authored-by: Craig Cook <craigcook@users.noreply.github.com>